Revert "[FIX] delivery, stock: carrier_id field in incoming stock picking"
This reverts commit 8fbcfda9e7e9c26d997db814d0fdb2bee20a0414.

This was no longer possible to create a picking from scratch.
`Inventory > Your Company: Receipts > Create`
Because `picking_type_id` is not in the res returned by the above call to default_get

Even by forcing `picking_type_id` in the `fields` list,
it doesn't return the value in the fields.