[IMP] project_issue: remove 'team_id' on project issue + code cleaning
The sales team defined on the project issue is never used.

Furtermore, in the method 'stage_find' contains a parameter team_id BUT
a project_id is given when called.

That way, we can also remove the dependency to sales_team which doesn't make
sense anymore, and remove 2 useless ir.model.access rules