[REV] Packaging: remove "openerp-gevent" from script
This reverts commit 876098b18b8de3992b220229749f76608cb9723d and thus
allow to actually use the longpolling feature.

The commit was made originally to better follow the debian packaging
guidelines.

A better solution will be implemented for v10: removing the `openerp-gevent`
script to replace it with a special argument in `odoo.py`.

There is still a dependency issue as we need psycogreen but this dependency
is not met in debian wheezy. This will be made a hard dependency for v10 as
it is available in debian jessie.

opw 678334