[ADD] base: add company_name field on res_partner
res_partner now have a new concept of company_name. (duplicaed of parent_id)

That allow to a frontend user to have a company name without create a useless
tree of "Company -> Partner -> Contacts"

Before this commit, we was using street2 or street field to do the same thing.
That was confusing in some screens because according from where the partner has
been created, the company name was stored in distinct field and that was like
a bug for some users.

commercial_company_name is based on the same behavior of commercial_partner_id.
It is base on the field company_name.

In Partner form wiew, a quick button has been added in edit mode if however
the user want to create the partner tree to add new collegues.