[FIX] hr_attendance: Attendance Error Report fail afer upgrade
After upgrading the `hr_attendance` module,
it was no longer possible to use the
`Attendance Error Report` link in the `More`
menu of employees.

This is because there were two binding actions
added in the code under the samename
(`Attendance Error Report`):
 - `hr_attendance_error_values`
 - `action_report_hrattendanceerror`

The first launch a wizard which will print the report
after filling the required data,
The second tries to directly print the report, but it
fails because required data (from the wizard) is missing.

Before update, it's the first binding action that was displayed.
After update, it was the second, which failed.

We solve this issue by changing the way the report is
created through the data, by not using a `<report/>` tag,
this one creating automatically the action binding.
Instead, we use a classic `<record/>` tag, which
doesn't create automatically the action binding.

opw-682720