[FIX] account: unnecessary default currency
A mechanism already exists to select the appropriate currency and was
introduced in commit 27d3ab9. However, it requires a purchase journal in
the currency of the PO to work.

Note that this commit is likely to not work correctly since the method
`_onchange_journal_id` is triggered after `_default_currency`.
Therefore, even if the currency is correctly set in `_default_currency`,
it is changed to the journal currency right after.

This reverts commit 9eeed9a3bb5c95c3805f44fa23ae61290528dd43.