[FIX] mass_mailing: do not inherit a view on a `string` attribute
A string attribute value can be translated,
and therefore the inherit broken if the term is indeed
translated.

This was the case for this term "Tracked link" in the Dutch
language, the "Tracked link" term of the inherited view has
been translated, and the inherit was therefore broken as
it couldn't find the given node in the base view due to the
fact this term was translated.

opw-683128