[FIX] sale_timesheet: don't count expenses as timesheets
On sale orders, the timesheets stat button should display the number of related timesheets.
a52492effb305e560a6b3743e8495b3f73dd3082 changed the computation system to have the number of timesheets
instead of the total time for several reasons, but used a domain which also includes expenses.

This commit changes the domain to make sure that the stat button only displays the number of timesheets,
and that the action leads to a list view of timesheets only.