[FIX] account: return correct values for default call
In the new API, default calls are supposed to return recordset
(even empty recordset), not bool values for relationnal fields.

If the table for this model was already present and if no l10n_x
was installed, no default value could be computed correctly for
existing entries. This caused a crash on install.

opw-683470