[IMP] account/stock_account: Add a clickable history on journal entries chatters
Purpose:
In a perpetual inventory mode, it is not possible to reach the picking from the journal entry view because the ref field is not clickable. I propose to show a chatter to get the history (which doc created the move) and being able to go to the move from the chatter.

Specification:
In journal entry view add chatter and generate message(clickable) that link to  its Origin (i.e Supplier Invoice , Customer Invoice, Picking in case of Perpetual Inventory)