[IMP] web: logout client action: prevent toggling app switcher
When the webclient calls do_action from the app_switcher, it toggles it when
the returned deferred is resolved.

The logout client action didn't return anything, so when logging out from the
app switcher, the webclient directly toggled the app_switcher and displayed
the last opened action if any, and a gray screen otherwise.

This commit makes the client action return a deferred which is never resolved,
so the app_switcher isn't toggled when logging out.