[FIX] web: session_info: introduce `is_superuser` key
This is done in order to enable the tour service only for the
superuser. Commit e45e2ae9f9e8c5 changed the meaning of `is_admin`
to achieve this but broke the fact that some features of the debug
manager are available to admin and not only to the superuser. Anyway
a `is_admin` key should mean that the user is in the admin group.

We could have checked client side that the uid is equal to 1 but
adding a `is_superuser` key is more elegant and easier to grep.
Also fix the only occurence of a uid check in the weblcient.

This reverts commit e45e2ae9f9e8c536991a8c7894fbb32201f53410.